Sync the runtime and generated parity tests through a shared scenario base
EngineParityTests defines each parity scenario once; the RuntimeConverterParityTests
(builders) and GeneratedConverterParityTests (generated converters) fixtures only
supply their options. The shared domain types live in ParityDomain.cs. 17 scenarios
now run against both engines (34 tests), so a behavioral fix in one engine is
automatically checked against the other. Engine-specific divergences (nested
serialization, attribute-based base-as-leaf write) stay in the generated fixture.
